backend_flask/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
Flask
flask-cors
python-dotenv
torch==2.2.2+cpu
torchvision==0.17.2+cpu
facenet-pytorch
numpy
scikit-learn
Pillow